device-telemetry-alerts/services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
psycopg2-binary==2.9.9
redis==5.0.8
python-jose[cryptography]==3.3.0
prometheus-client==0.20.0
boto3==1.34.162
